backend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
pathspec==0.11.2
python-magic>=0.4.27
pypdf>=4.1.0
python-docx>=1.1.0
striprtf>=0.0.26
textract>=1.6.5
spacy>=3.7.4
# Production dependencies
psycopg2-binary>=2.9.0
dj-database-url>=2.1.0
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0